fix reports for areas without id
